moved extensions to `root`
