+++ ceres-ads/api/ceres-ads.api
@@ -59,6 +59,10 @@ public final class dev/teogor/ceres/ads/AdView : androidx/appcompat/widget/AppCo
 	public final fun setColor (I)V
 }
 
+public final class dev/teogor/ceres/ads/AdsExt {
+	public static final fun isAdActivity (Landroid/app/Activity;)Z
+}
+
 public final class dev/teogor/ceres/ads/AdsModule : dev/teogor/ceres/core/app/ModuleProvider {
 	public fun <init> ()V
 	public fun onCreate ()V
@@ -241,10 +245,6 @@ public final class dev/teogor/ceres/ads/utils/Constants$TestAdsId {
 	public static final field NATIVE_VIDEO Ljava/lang/String;
 	public static final field REWARDED Ljava/lang/String;
 	public static final field REWARDED_INTERSTITIAL Ljava/lang/String;
-}
-
-public final class dev/teogor/ceres/ads/utils/ExtensionsKt {
-	public static final fun isAdActivity (Landroid/app/Activity;)Z
 }
 
 public final class dev/teogor/ceres/ads/utils/SystemKt {
